move configuration functions to new file
