+++ beacon-chain/node/BUILD.bazel
@@ -4,6 +4,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_test")
 go_library(
     name = "go_default_library",
     srcs = [
+        "configuration.go",
         "log.go",
         "node.go",
     ],
